quest-narrative-designer/quest-narrative-designer/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0
openai>=0.28.0
python-dotenv>=1.0.0
